New feature: tag navigation in refactor menu
Enable navigation with "n" and "p" keys to move to the semantic unit in
code buffer when cursor is in a menu item of a tag in menu buffer. Also
add overlay to highlight the tag position in code buf.ref

Add keymap slot that is a function that set keys
srefactor-ui-menu-mode-map.