Handle tag moving with comment and macro
Function can have macro at the beginning, and sometimes Semantic cannot
recognize it. Another thing is that moving tag without moving the
comment defeats the purpose of moving tag because user has to move back
to cut the comment anyway.

Add srefactor--tag-start-with-comment function to handle these cases. It
should only be used ONCE to retrieve the beginning position and set that
position to a tag. Do not feed the same tag again into the function,
otherwise incorrect position is returned.