Fix local variable rename for shorter new name
When the new name is shorter, after the first name is renamed, the other
positions in the list are no longer positions at the beginning of line.
We need those positions less than the position of the first character of
each occurrence for search-forward-regexp to work.

This change moves each point to beginning of line to avoid this
situation. However, we need to fix this properly using mark.