Handle cases where an sexp needs newlines between its sub-sexpressions
For example, everything inside "cond" pretty much has a newline between
sexpressions, but it's not the case for let or let*. With current
mechanism, every lexical token (except the parenthesis) stands on it onw
line and we make it pretty by delete indentation of the symbol after the
header symbol. However, this is not the case with sexp like "cond", so
we handle it specially by passing newline-between-semantic-lists.
Newline here simply means that we don't delete indentation of the token after
the head symbol to join it with the head symbol, thus creating "newline".