Finally handle symbol with number at front
Now I understand why most languages disallow numbers as the first
character of an identifier. Such ambiguity.

In this case, check if the next-next-token is a semantic list, and if so
insert a newline, so the next sexp won't join the same line. Also
remove the erroneous pairs ("1" . 2) in skip list. It should be ("1+"
. 2) and ("1-" . 2).

It's all about proper token skip.