fix: stop inserting when encounter parens
Even if ignore-num is still greater than 0, but when we encounter a
close-paren, it means that the form does not have enough arguments. If
we keep inserting, we insert extra parentheses because we satisfy the if
below and decrease the count of ignore-num by 1. Insetad, put it back to
the token stream and let punctuation formatter handle it.