[Fix #29] Don't stop when comment start not found
semantic-documentation-comment-preceeding-tag can incorrectly identify
comment preceeding a tag. Specifically, when there's an inline comment
preceeding a tag, and the comment belongs to a different line of
code (i.e. a macro), semantic-documentation-comment-preceeding-tag still
counts that comment as the comment of current tag.

Thus, the fix is to prevent search-backward-regexp to spit error then
stop the whole operation. Instead, fall back to the else case.

Also, add another check to ensure that the comment found by
search-backward-regexp really belongs to current tag by checking if the
comment is prefixed with only zero or more spaces. If the comment is
prefixed with code, then obviously it belongs to some other statement,
not current tag.